pyproject.toml
[project]
name = "aida"
version = "0.1.0"
description = "Aida - AI Desktop Assistant for KDE Plasma"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Ruudo" }]

dependencies = [
    "PySide6>=6.6.0",
    "faster-whisper>=1.0.0",
    "piper-tts>=1.2.0",
    "ollama>=0.3.0",
    "opencv-python>=4.9.0",
    "playwright>=1.40.0",
    "numpy>=1.26.0",
    "sounddevice>=0.4.6",
    "soundfile>=0.12.1",
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "sentence-transformers>=2.2.0",
    "faiss-cpu>=1.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-qt>=4.2.0",
    "ruff>=0.1.0",
]

[project.scripts]
aida = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
